Feature/j1 tests (#13)
tests for j1 programs

+++ src/s_expression.rs
@@ -2,7 +2,7 @@ use std::{iter::Peekable, str::Chars};
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub enum SExpr {
-    Num(i64),
+    Num(i32),
     Sym(String),
     List(Vec<SExpr>),
 }
@@ -43,7 +43,7 @@ fn skip_white_space(it: &mut Peekable<Chars>) {
 
 // should return a list
 fn read_list(it: &mut Peekable<Chars>) -> SExpr {
-    // ( + e e )
+    // (+ e e)
     it.next(); // consume + 2 3 )
     let mut elements = Vec::new();
     loop {
@@ -67,7 +67,7 @@ fn read_list(it: &mut Peekable<Chars>) -> SExpr {
 
 
 // A number or a symbol 
-// ( + 2 3 ) put everything inside the () into a
+// (+ 2 3) put everything inside the () into a
 fn read_atom(it: &mut Peekable<Chars>) -> SExpr {
     let mut s = String::new();
 
@@ -78,7 +78,7 @@ fn read_atom(it: &mut Peekable<Chars>) -> SExpr {
         s.push(it.next().unwrap());
     }
 
-    if let Ok(n) = s.parse::<i64>() {
+    if let Ok(n) = s.parse::<i32>() {
         SExpr::Num(n)
     } else {
         SExpr::Sym(s)
